feat: fix import, remove unused inspector props
